Vision/PID/diagramskill/requirements.txt
azure-functions
opencv-contrib-python==4.2.0.34  
opencv-python==4.2.0.34 
requests
Pillow
azure-storage-blob
pytesseract
datetime